requirements.txt
Django~=3.1.2
django_filters~=2.3.0
django-bootstrap4~=2.2.0
django-bootstrap-datepicker-plus~=3.0.5